Create TiledImageViewPrintPageRenderer.m
Will properly size and paginate for printing an Image obtained from the TiledImageView on iOS. If Contents is suited to Landscape it will paginate to print pages in a row left to right. If content is suited to portrait it will paginate to print pages in column. This can easily be modified to suit your needs if different.